Rename package from ApiAlerts.Common to APIAlerts
 - Client renamed to service
 - Main class is now Client
 - Alert object
 - Added base Script.GitHub to test project
