+++ ApiAlerts.sln
@@ -3,9 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34601.278
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ApiAlerts.Common", "src\ApiAlerts.Common\ApiAlerts.Common.csproj", "{DB9AA57B-BCDB-4229-B562-E9AAF493323B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "APIAlerts", "src\APIAlerts\APIAlerts.csproj", "{DB9AA57B-BCDB-4229-B562-E9AAF493323B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ApiAlerts.Common.Tests", "tests\ApiAlerts.Common.Tests\ApiAlerts.Common.Tests.csproj", "{9A20F2E2-5CAC-443A-AC48-8A8A6456076B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "APIAlerts.Tests", "tests\APIAlerts.Tests\APIAlerts.Tests.csproj", "{9A20F2E2-5CAC-443A-AC48-8A8A6456076B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Script.GitHub", "Script.GitHub\Script.GitHub.csproj", "{5C3FC72A-9F66-4213-BFF1-B43DDF8D29AD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{9A20F2E2-5CAC-443A-AC48-8A8A6456076B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9A20F2E2-5CAC-443A-AC48-8A8A6456076B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9A20F2E2-5CAC-443A-AC48-8A8A6456076B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5C3FC72A-9F66-4213-BFF1-B43DDF8D29AD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5C3FC72A-9F66-4213-BFF1-B43DDF8D29AD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5C3FC72A-9F66-4213-BFF1-B43DDF8D29AD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5C3FC72A-9F66-4213-BFF1-B43DDF8D29AD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
