+++ src/APIAlerts/Constants.cs
@@ -1,0 +1,8 @@
+namespace APIAlerts;
+
+internal static class Constants
+{
+    internal const string BaseUrl = "https://api.apialerts.com";
+    internal const string IntegrationName = "dotnet";
+    internal const string Version = "1.1.0";
+}
