+++ src/APIAlerts/network/contract/ErrorResponse.cs
@@ -1,0 +1,9 @@
+using System.Text.Json.Serialization;
+
+namespace APIAlerts.network.contract;
+
+internal class ErrorResponse
+{
+    [JsonPropertyName("message")]
+    public string? Message { get; set; }
+}
