--- src/ApiAlerts.Common/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("ApiAlerts.Common.Tests")]
