--- src/ApiAlerts.Common/Constants.cs
@@ -1,8 +1,0 @@
-namespace ApiAlerts.Common;
-
-internal static class Constants
-{
-    internal const string BaseUrl = "https://api.apialerts.com";
-    internal const string IntegrationName = "dotnet";
-    internal const string Version = "1.1.0";
-}
