--- src/ApiAlerts.Common/network/contract/ErrorResponse.cs
@@ -1,9 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace ApiAlerts.Common.network.contract;
-
-internal class ErrorResponse
-{
-    [JsonPropertyName("message")]
-    public string? Message { get; set; }
-}
